added validation and erorr handling

+++ src/routes/users.js
@@ -1,5 +1,8 @@
 import { Router } from 'express'
 import UserModal from '../db-models/userSchema.js'
+import { userCreationValidator } from '../middleware/validation.js'
+import { validationResult } from 'express-validator'
+import createHttpError from 'http-errors'
 
 const userRouter = Router()
 
@@ -12,8 +15,10 @@ userRouter.route('/')
         next(error)
     }
 })
-.post(async (req, res, next) => {
+.post(userCreationValidator, async (req, res, next) => {
     try {
+        const errors = validationResult(req)
+        if (!errors.isEmpty()) return next(createHttpError(400, errors))
         const newUser = new UserModal(req.body)
         newUser.save()
         res.send(newUser)
@@ -25,7 +30,9 @@ userRouter.route('/')
 userRouter.route('/:userId')
 .get(async (req, res, next) => {
     try {
+        if (req.params.userId.length !== 24) return next(createHttpError(400, 'Invalid ID'))
         const user = await UserModal.findById(req.params.userId)
+        if (!user) return next(createHttpError(400, `The id ${req.params.userId} does not match any users`))
         res.send(user)
     } catch (error) {
         console.log(error)
@@ -34,7 +41,9 @@ userRouter.route('/:userId')
 })
 .put(async (req, res, next) => {
     try {
+        if (req.params.userId.length !== 24) return next(createHttpError(400, 'Invalid ID'))
         const updatedUser = await UserModal.findByIdAndUpdate(req.params.userId, req.body, { new: true })
+        if (!updatedUser) return next(createHttpError(400, `The id ${req.params.userId} does not match any users`))
         res.send(updatedUser)
     } catch (error) {
         next(error)
@@ -42,7 +51,9 @@ userRouter.route('/:userId')
 })
 .delete(async (req, res, next) => {
     try {
+        if (req.params.userId.length !== 24) return next(createHttpError(400, 'Invalid ID'))
         const result = await UserModal.findByIdAndDelete(req.params.userId)
+        if (!result) return next(createHttpError(400, `The id ${req.params.userId} does not match any users`))
         res.status(204).send()
     } catch (error) {
         next(error)
